requirements.txt
Django==1.2.1
Fabric==0.9.1
django-nose==0.1
